+++ rotkehlchen/chain/manager.py
@@ -1248,3 +1248,24 @@ class ChainManager(CacheableObject, LockableQueryObject):
                 ))
 
         return defi_events
+
+    @cache_response_timewise()
+    def get_loopring_balances(self) -> Dict[Asset, Balance]:
+        """Query loopring balances if the module is activated"""
+        # Check if the loopring module is activated
+        loopring_module = self.get_module('loopring')
+
+        if loopring_module is None:
+            return {}
+
+        addresses = self.queried_addresses_for_module('loopring')
+        balances = loopring_module.get_balances(addresses=addresses)
+
+        # Now that we have balances for the addresses we need to aggregate the
+        # assets in the different addresses
+        aggregated_balances: Dict[Asset, Balance] = defaultdict(Balance)
+        for _, assets in balances.items():
+            for asset, balance in assets.items():
+                aggregated_balances[asset] += balance
+
+        return dict(aggregated_balances)