+++ rotkehlchen/serialization/deserialize.py
@@ -477,6 +477,8 @@ def deserialize_location(symbol: str) -> Location:
         return Location.KUCOIN
     if symbol == 'balancer':
         return Location.BALANCER
+    if symbol == 'loopring':
+        return Location.LOOPRING
     # else
     raise DeserializationError(
         f'Failed to deserialize location symbol. Unknown symbol {symbol} for location',
@@ -590,6 +592,8 @@ def deserialize_location_from_db(symbol: str) -> Location:
         return Location.KUCOIN
     if symbol == 'X':
         return Location.BALANCER
+    if symbol == 'Y':
+        return Location.LOOPRING
     # else
     raise DeserializationError(
         f'Failed to deserialize location symbol. Unknown symbol {symbol} for location',